fix(libchkr): do not ouput the name of the file with 'file' command.
Files whose name contains the word 'directory' were considered as directories
since the name of the file was in the ouput of the 'file' command.
We now use the --brief option so that only the type of the file is in the
output and the name of the file does not interfere with type comparison.